app/requirements.txt
pandasai
streamlit
rdkit
bs4
selenium
duckdb
duckdb-engine